add condition to make sure form isn't loaded until friends are retrieved
